refactor(scripts): move alert and peg validators into scripts/alerts/ (P8)
Phase P8 of the scripts/ reorganization (issue 1877). Moves the eight files of
the alert-rule and peg-policy cluster into scripts/alerts/ and updates every
caller in the same PR. check-peg-policy-publication was filed under the tf
cluster only because tf-stacks.test.mjs side-effect imports its test; it is a
peg validator, and the move costs that umbrella one path segment.

pegPolicyVersionDigest existed twice, byte-identical, in alert-rules-lint.mjs
and check-peg-registry-integrity.mjs. Both now read one definition in
scripts/lib/peg-policy-digest.mjs, which also owns the canonical fingerprint the
lineage comparison comes down to and the digest-suffix pattern that reads what
the digest writes.

Two files split at their real seams. alert-rules-lint.mjs (1,049 lines) keeps
the gauge cross-check and the run; extraction, lezer parsing, and the
ALERT_RULES_LINT_MIN_* floors move to alert-rules-lint-extract.mjs, and the peg
bundle and selector rules move to alert-rules-lint-peg-policy.mjs. The spec for
this phase described the second half as the metrics-bridge gauge cross-check;
in the source that check is under 30 lines and the 450-line half is peg policy.
check-peg-registry-integrity.mjs (1,042 lines) sheds its git-shelling policy
lineage comparison to check-peg-registry-integrity-lineage.mjs. Every moved line
is byte-identical to its source.

check-peg-registry-integrity.mjs resolves @mento-protocol/config through a
two-base createRequire because pnpm does not hoist that metrics-bridge
dependency to the workspace root. Both bases are relative to the file, so the
second is re-anchored to ../../metrics-bridge/package.json. Proven at the new
depth: the first base still fails MODULE_NOT_FOUND, the second still resolves,
and the stale ../metrics-bridge/package.json fails.

Swept: the alias pair in package.json and the pinned alias map, four enumerated
paths-filter blocks across ci.yml, infra.yml, and peg-policy-publication.yml,
seven workflow run lines, terraform.stacks.json, workflow-inventory.mjs, the
tf-stacks umbrella import, three ADRs, and two source comments.
Splitting code out of a listed file narrows what that file's entry covers, so
the modules the two peg checkers import are listed alongside them:
peg-policy-publication.yml and terraform.stacks.json gain all three, hcl.mjs
included, and the ci.yml and infra.yml terraform filters gain the two this PR
creates. P7 set that precedent when it added hcl.mjs and workflow-yaml.mjs to
the same three filters.

The gate carries fifteen literal command and arm strings for these files, six
of them outside the focused arms. All are repointed, the split modules join the
arms that read them, and scripts/lib/peg-policy-digest.mjs gets its own arm
running both peg suites. The publication boundary gains the focused arm it never
had; the unconditional sweep already ran its suite through pnpm tf:test, so this
only makes the reason honest. The gate test gains routing assertions for all six
paths.

scripts/AGENTS.md records the landed phase, the lib/ tier's third member, and
why redrive-onchain-deadletter stays flat. It was 63 bytes under its scoped cap,
so the tf:test operating rule now routes to ADR 0053 and docs/deployment.md,
which already carry that detail verbatim, instead of restating it.
