+++ .github/workflows/peg-policy-publication.yml
@@ -20,8 +20,15 @@ on:
       - alerts/peg-policy-publication/**
       - alerts/rules/peg-thresholds.json
       - metrics-bridge/peg-registry.json
-      - scripts/check-peg-policy-publication.mjs
-      - scripts/check-peg-registry-integrity.mjs
+      - scripts/alerts/check-peg-policy-publication.mjs
+      - scripts/alerts/check-peg-registry-integrity.mjs
+      # Modules the two checkers above import (ADR 0064). They sit outside both
+      # entrypoints, so without their own entries a change to the lineage
+      # comparison, the version digest, or the HCL parser never fires this
+      # validation.
+      - scripts/alerts/check-peg-registry-integrity-lineage.mjs
+      - scripts/lib/peg-policy-digest.mjs
+      - scripts/lib/hcl.mjs
       - .github/workflows/peg-policy-publication.yml
   # Deliberately no push trigger: policy publication stays an explicit,
   # production-infra-gated operation after its documented rollout gates.
@@ -68,8 +75,8 @@ jobs:
       - name: Validate Peg registry and publication boundary
         working-directory: ${{ github.workspace }}
         run: |
-          node scripts/check-peg-registry-integrity.mjs
-          node scripts/check-peg-policy-publication.mjs
+          node scripts/alerts/check-peg-registry-integrity.mjs
+          node scripts/alerts/check-peg-policy-publication.mjs
 
   plan:
     # autofix-ci-trust: this job only runs for manual main dispatches, never a
