+++ package.json
@@ -107,8 +107,8 @@
     "infra:init": "terraform -chdir=terraform init",
     "infra:plan": "pnpm tf plan platform",
     "alerts:rules:init": "terraform -chdir=alerts/rules init",
-    "alerts:rules:lint": "node scripts/alert-rules-lint.mjs",
-    "alerts:rules:lint:test": "node scripts/alert-rules-lint.test.mjs",
+    "alerts:rules:lint": "node scripts/alerts/alert-rules-lint.mjs",
+    "alerts:rules:lint:test": "node scripts/alerts/alert-rules-lint.test.mjs",
     "alerts:rules:plan": "pnpm tf plan alerts-rules",
     "alerts:infra:init": "terraform -chdir=alerts/infra init",
     "alerts:infra:plan": "pnpm tf plan alerts-delivery",
