+++ scripts/alerts/check-peg-policy-publication.mjs
@@ -11,9 +11,13 @@ import {
   nestedBlocks,
   terraformTopLevelBlocks,
   topLevelBlockKey,
-} from "./lib/hcl.mjs";
+} from "../lib/hcl.mjs";
 
-const root = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
+const root = path.resolve(
+  path.dirname(fileURLToPath(import.meta.url)),
+  "..",
+  "..",
+);
 export const PUBLICATION_FILES = [
   "versions.tf",
   "providers.tf",
