+++ terraform.stacks.json
@@ -63,9 +63,12 @@
         "terraform/peg-policy.tf",
         "terraform/ci-wif.tf",
         "terraform.stacks.json",
-        "scripts/check-peg-policy-publication.mjs",
-        "scripts/check-peg-policy-publication.test.mjs",
-        "scripts/check-peg-registry-integrity.mjs",
+        "scripts/alerts/check-peg-policy-publication.mjs",
+        "scripts/alerts/check-peg-policy-publication.test.mjs",
+        "scripts/alerts/check-peg-registry-integrity.mjs",
+        "scripts/alerts/check-peg-registry-integrity-lineage.mjs",
+        "scripts/lib/peg-policy-digest.mjs",
+        "scripts/lib/hcl.mjs",
         "scripts/production-infra-identity-contract/**",
         "scripts/terraform-fmt-check.mjs",
         "scripts/tf-stacks.mjs",
